+++ test/system/distr_tests.cc
@@ -388,12 +388,15 @@ namespace detail {
 		auto& tm = celerity::detail::runtime::get_instance().get_task_manager();
 		tm.set_horizon_step(1);
 
-		for(int i = 0; i < 2; ++i) {
-			q.submit([&](handler& cgh) {
-				celerity::accessor acc_a{buff_a, cgh, celerity::access::one_to_one{}, celerity::read_write};
-				cgh.parallel_for(range, [=](item<2> item) { (void)acc_a; });
-			});
-		}
+		q.submit([&](handler& cgh) {
+			celerity::accessor acc_a{buff_a, cgh, celerity::access::one_to_one{}, celerity::write_only, celerity::no_init};
+			cgh.parallel_for(range, [=](item<2> item) { (void)acc_a; });
+		});
+
+		q.submit([&](handler& cgh) {
+			celerity::accessor acc_a{buff_a, cgh, celerity::access::one_to_one{}, celerity::read_write};
+			cgh.parallel_for(range, [=](item<2> item) { (void)acc_a; });
+		});
 
 		q.slow_full_sync();
 
@@ -432,6 +435,45 @@ namespace detail {
 			CHECK(dot == expected);
 			if(dot != expected) { fmt::print("\n{}:\n\ngot:\n\n{}\n\nexpected:\n\n{}\n\n", Catch::getResultCapture().getCurrentTestName(), dot, expected); }
 		}
+	}
+
+	TEST_CASE_METHOD(test_utils::runtime_fixture, "runtime logs errors on overlapping writes between commands iff access pattern diagnostics are enabled",
+	    "[runtime][diagnostics]") //
+	{
+		std::unique_ptr<celerity::test_utils::log_capture> lc;
+		{
+			distr_queue q;
+			const auto num_nodes = runtime::get_instance().get_num_nodes();
+			if(num_nodes < 2) { SKIP("Test needs at least 2 participating nodes"); }
+
+			lc = std::make_unique<celerity::test_utils::log_capture>();
+
+			buffer<int, 1> buf(1);
+
+			SECTION("in distributed device kernels") {
+				q.submit([&](handler& cgh) {
+					accessor acc(buf, cgh, celerity::access::all(), write_only, no_init);
+					cgh.parallel_for(range(num_nodes), [=](item<1>) { (void)acc; });
+				});
+			}
+
+			SECTION("in collective host tasks") {
+				q.submit([&](handler& cgh) {
+					accessor acc(buf, cgh, celerity::access::all(), write_only_host_task, no_init);
+					cgh.host_task(celerity::experimental::collective, [=](experimental::collective_partition) { (void)acc; });
+				});
+			}
+
+			q.slow_full_sync();
+		}
+
+		const auto error_message = "has overlapping writes between multiple nodes in B0 {[0,0,0] - [1,1,1]}. Choose a non-overlapping range mapper for the "
+		                           "write access or constrain the split to make the access non-overlapping.";
+#if CELERITY_ACCESS_PATTERN_DIAGNOSTICS
+		CHECK_THAT(lc->get_log(), Catch::Matchers::ContainsSubstring(error_message));
+#else
+		CHECK_THAT(lc->get_log(), !Catch::Matchers::ContainsSubstring(error_message));
+#endif
 	}
 
 } // namespace detail